Buyables/creatables changes, max notifications (#2058)
* buyables cleanup/refactoring

* commit

* move around creatables

* add max capes to cl and creatables

* fix max capes cl

* add max notification

* use event notif

* fix errors

* fix err
