+++ src/lib/data/buyables/canifisClothes.ts
@@ -4,185 +4,116 @@ import { Buyable } from './buyables';
 export const canifisClothes: Buyable[] = [
 	{
 		name: 'Grey gloves',
-		outputItems: resolveNameBank({
-			'Grey gloves': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Grey boots',
-		outputItems: resolveNameBank({
-			'Grey boots': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Grey robe bottoms',
-		outputItems: resolveNameBank({
-			'Grey robe bottoms': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Grey hat',
-		outputItems: resolveNameBank({
-			'Grey boots': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Grey robe top',
-		outputItems: resolveNameBank({
-			'Grey robe top': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Red gloves',
-		outputItems: resolveNameBank({
-			'Red gloves': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Red boots',
-		outputItems: resolveNameBank({
-			'Red boots': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Red robe bottoms',
-		outputItems: resolveNameBank({
-			'Red robe bottoms': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Red robe top',
-		outputItems: resolveNameBank({
-			'Red robe top': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Red hat',
-		outputItems: resolveNameBank({
-			'Red hat': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Yellow gloves',
-		outputItems: resolveNameBank({
-			'Yellow gloves': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Yellow boots',
-		outputItems: resolveNameBank({
-			'Yellow boots': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Yellow robe top',
-		outputItems: resolveNameBank({
-			'Yellow robe top': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Yellow robe bottoms',
-		outputItems: resolveNameBank({
-			'Yellow robe bottoms': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Yellow hat',
-		outputItems: resolveNameBank({
-			'Yellow hat': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Teal gloves',
-		outputItems: resolveNameBank({
-			'Teal gloves': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Teal boots',
-		outputItems: resolveNameBank({
-			'Teal boots': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Teal robe top',
-		outputItems: resolveNameBank({
-			'Teal robe top': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Teal robe bottoms',
-		outputItems: resolveNameBank({
-			'Teal robe bottoms': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Teal hat',
-		outputItems: resolveNameBank({
-			'Teal hat': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Purple gloves',
-		outputItems: resolveNameBank({
-			'Purple gloves': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Purple boots',
-		outputItems: resolveNameBank({
-			'Purple boots': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Purple robe top',
-		outputItems: resolveNameBank({
-			'Purple robe top': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
@@ -196,49 +127,31 @@ export const canifisClothes: Buyable[] = [
 	},
 	{
 		name: 'Purple hat',
-		outputItems: resolveNameBank({
-			'Purple hat': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Red cape',
-		outputItems: resolveNameBank({
-			'Red cape': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Black cape',
-		outputItems: resolveNameBank({
-			'Black cape': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Blue cape',
-		outputItems: resolveNameBank({
-			'Blue cape': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Yellow cape',
-		outputItems: resolveNameBank({
-			'Yellow cape': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	},
 	{
 		name: 'Green cape',
-		outputItems: resolveNameBank({
-			'Green cape': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 15
 	}
