+++ src/lib/data/buyables/frem.ts
@@ -1,92 +1,58 @@
-import { resolveNameBank } from '../../util';
 import { Buyable } from './buyables';
 
 export const fremennikClothes: Buyable[] = [
 	{
 		name: 'Fremennik green cloak',
-		outputItems: resolveNameBank({
-			'Fremennik green cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik blue cloak',
-		outputItems: resolveNameBank({
-			'Fremennik blue cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik brown cloak',
-		outputItems: resolveNameBank({
-			'Fremennik brown cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik cyan cloak',
-		outputItems: resolveNameBank({
-			'Fremennik cyan cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik red cloak',
-		outputItems: resolveNameBank({
-			'Fremennik red cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik grey cloak',
-		outputItems: resolveNameBank({
-			'Fremennik grey cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik yellow cloak',
-		outputItems: resolveNameBank({
-			'Fremennik yellow cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik teal cloak',
-		outputItems: resolveNameBank({
-			'Fremennik teal cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik purple cloak',
-		outputItems: resolveNameBank({
-			'Fremennik purple cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik pink cloak',
-		outputItems: resolveNameBank({
-			'Fremennik pink cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	},
 	{
 		name: 'Fremennik black cloak',
-		outputItems: resolveNameBank({
-			'Fremennik black cloak': 1
-		}),
 		gpCost: 100_000,
 		qpRequired: 55
 	}
