+++ src/lib/data/createables.ts
@@ -3,8 +3,11 @@ import { ItemBank } from '../types';
 import { itemNameFromID, resolveNameBank } from '../util';
 import itemID from '../util/itemID';
 import { metamorphPets } from './collectionLog';
+import { capeCreatables } from './creatables/capes';
+import { ornamentKits } from './creatables/ornaments';
+import { slayerCreatables } from './creatables/slayer';
 
-interface Createable {
+export interface Createable {
 	name: string;
 	outputItems: ItemBank;
 	inputItems: ItemBank;
@@ -159,273 +162,6 @@ const crystalTools: Createable[] = [
 	}
 ];
 
-const ornamentKits: Createable[] = [
-	{
-		name: 'Dragon defender (t)',
-		inputItems: resolveNameBank({ 'Dragon defender': 1, 'Dragon defender ornament kit': 1 }),
-		outputItems: resolveNameBank({ 'Dragon defender (t)': 1 })
-	},
-	{
-		name: 'Dragon defender',
-		inputItems: resolveNameBank({ 'Dragon defender (t)': 1 }),
-		outputItems: resolveNameBank({ 'Dragon defender': 1, 'Dragon defender ornament kit': 1 }),
-		noCl: true
-	},
-	{
-		name: 'Rune defender (t)',
-		inputItems: resolveNameBank({ 'Rune defender': 1, 'Rune defender ornament kit': 1 }),
-		outputItems: resolveNameBank({ 'Rune defender (t)': 1 })
-	},
-	{
-		name: 'Rune defender',
-		inputItems: resolveNameBank({ 'Rune defender (t)': 1 }),
-		outputItems: resolveNameBank({ 'Rune defender': 1, 'Rune defender ornament kit': 1 }),
-		noCl: true
-	},
-	{
-		name: 'Dragon pickaxe (or)',
-		inputItems: resolveNameBank({ 'Dragon pickaxe': 1, 'Zalcano shard': 1 }),
-		outputItems: resolveNameBank({ 'Dragon pickaxe (or)': 1 })
-	},
-	{
-		name: 'Dragon pickaxe',
-		inputItems: resolveNameBank({ 'Dragon pickaxe (or)': 1 }),
-		outputItems: resolveNameBank({ 'Dragon pickaxe': 1, 'Zalcano shard': 1 }),
-		noCl: true
-	},
-	{
-		name: 'Dragon sq shield (g)',
-		inputItems: resolveNameBank({ 'Dragon sq shield ornament kit': 1, 'Dragon sq shield': 1 }),
-		outputItems: resolveNameBank({ 'Dragon sq shield (g)': 1 })
-	},
-	{
-		name: 'Revert dragon sq shield',
-		inputItems: resolveNameBank({ 'Dragon sq shield (g)': 1 }),
-		outputItems: resolveNameBank({ 'Dragon sq shield ornament kit': 1, 'Dragon sq shield': 1 }),
-		noCl: true
-	},
-	{
-		name: 'Dragon platelegs (g)',
-		inputItems: resolveNameBank({ 'Dragon platelegs': 1, 'Dragon legs/skirt ornament kit': 1 }),
-		outputItems: resolveNameBank({ 'Dragon platelegs (g)': 1 })
-	},
-	{
-		name: 'Dragon platelegs',
-		inputItems: resolveNameBank({ 'Dragon platelegs (g)': 1 }),
-		outputItems: resolveNameBank({
-			'Dragon platelegs': 1,
-			'Dragon legs/skirt ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Dragon plateskirt (g)',
-		inputItems: resolveNameBank({
-			'Dragon plateskirt': 1,
-			'Dragon legs/skirt ornament kit': 1
-		}),
-		outputItems: resolveNameBank({ 'Dragon plateskirt (g)': 1 })
-	},
-	{
-		name: 'Dragon plateskirt',
-		inputItems: resolveNameBank({ 'Dragon plateskirt (g)': 1 }),
-		outputItems: resolveNameBank({
-			'Dragon plateskirt': 1,
-			'Dragon legs/skirt ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Dragon chainbody (g)',
-		inputItems: resolveNameBank({
-			'Dragon chainbody': 1,
-			'Dragon chainbody ornament kit': 1
-		}),
-		outputItems: resolveNameBank({ 'Dragon chainbody (g)': 1 })
-	},
-	{
-		name: 'Dragon chainbody',
-		inputItems: resolveNameBank({ 'Dragon chainbody (g)': 1 }),
-		outputItems: resolveNameBank({
-			'Dragon chainbody': 1,
-			'Dragon chainbody ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Amulet of fury (or)',
-		inputItems: resolveNameBank({ 'Amulet of fury': 1, 'Fury ornament kit': 1 }),
-		outputItems: resolveNameBank({
-			'Amulet of fury (or)': 1
-		})
-	},
-	{
-		name: 'Amulet of fury',
-		inputItems: resolveNameBank({
-			'Amulet of fury (or)': 1
-		}),
-		outputItems: resolveNameBank({ 'Amulet of fury': 1, 'Fury ornament kit': 1 }),
-		noCl: true
-	},
-	// Godswords
-	{
-		name: 'Zamorak godsword (or)',
-		inputItems: resolveNameBank({ 'Zamorak godsword': 1, 'Zamorak godsword ornament kit': 1 }),
-		outputItems: resolveNameBank({
-			'Zamorak godsword (or)': 1
-		})
-	},
-	{
-		name: 'Revert zamorak godsword',
-		inputItems: resolveNameBank({
-			'Zamorak godsword (or)': 1
-		}),
-		outputItems: resolveNameBank({ 'Zamorak godsword': 1, 'Zamorak godsword ornament kit': 1 }),
-		noCl: true
-	},
-	{
-		name: 'Bandos godsword (or)',
-		inputItems: resolveNameBank({ 'Bandos godsword': 1, 'Bandos godsword ornament kit': 1 }),
-		outputItems: resolveNameBank({
-			'Bandos godsword (or)': 1
-		})
-	},
-	{
-		name: 'Revert bandos godsword',
-		inputItems: resolveNameBank({
-			'Bandos godsword (or)': 1
-		}),
-		outputItems: resolveNameBank({ 'Bandos godsword': 1, 'Bandos godsword ornament kit': 1 }),
-		noCl: true
-	},
-	{
-		name: 'Saradomin godsword (or)',
-		inputItems: resolveNameBank({
-			'Saradomin godsword': 1,
-			'Saradomin godsword ornament kit': 1
-		}),
-		outputItems: resolveNameBank({
-			'Saradomin godsword (or)': 1
-		})
-	},
-	{
-		name: 'Revert saradomin godsword',
-		inputItems: resolveNameBank({
-			'Saradomin godsword (or)': 1
-		}),
-		outputItems: resolveNameBank({
-			'Saradomin godsword': 1,
-			'Saradomin godsword ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Armadyl godsword (or)',
-		inputItems: resolveNameBank({
-			'Armadyl godsword': 1,
-			'Armadyl godsword ornament kit': 1
-		}),
-		outputItems: resolveNameBank({
-			'Armadyl godsword (or)': 1
-		})
-	},
-	{
-		name: 'Revert Armadyl godsword',
-		inputItems: resolveNameBank({
-			'Armadyl godsword (or)': 1
-		}),
-		outputItems: resolveNameBank({
-			'Armadyl godsword': 1,
-			'Armadyl godsword ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Amulet of torture (or)',
-		inputItems: resolveNameBank({
-			'Amulet of torture': 1,
-			'Torture ornament kit': 1
-		}),
-		outputItems: resolveNameBank({
-			'Amulet of torture (or)': 1
-		})
-	},
-	{
-		name: 'Revert amulet of torture',
-		inputItems: resolveNameBank({
-			'Amulet of torture (or)': 1
-		}),
-		outputItems: resolveNameBank({
-			'Amulet of torture': 1,
-			'Torture ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Necklace of anguish (or)',
-		inputItems: resolveNameBank({
-			'Necklace of anguish': 1,
-			'Anguish ornament kit': 1
-		}),
-		outputItems: resolveNameBank({
-			'Necklace of anguish (or)': 1
-		})
-	},
-	{
-		name: 'Revert necklace of anguish',
-		inputItems: resolveNameBank({
-			'Necklace of anguish (or)': 1
-		}),
-		outputItems: resolveNameBank({
-			'Necklace of anguish': 1,
-			'Anguish ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Tormented bracelet (or)',
-		inputItems: resolveNameBank({
-			'Tormented bracelet': 1,
-			'Tormented ornament kit': 1
-		}),
-		outputItems: resolveNameBank({
-			'Tormented bracelet (or)': 1
-		})
-	},
-	{
-		name: 'Revert tormented bracelet',
-		inputItems: resolveNameBank({
-			'Tormented bracelet (or)': 1
-		}),
-		outputItems: resolveNameBank({
-			'Tormented bracelet': 1,
-			'Tormented ornament kit': 1
-		}),
-		noCl: true
-	},
-	{
-		name: 'Occult necklace (or)',
-		inputItems: resolveNameBank({
-			'Occult necklace': 1,
-			'Occult ornament kit': 1
-		}),
-		outputItems: resolveNameBank({
-			'Occult necklace (or)': 1
-		})
-	},
-	{
-		name: 'Revert occult necklace',
-		inputItems: resolveNameBank({
-			'Occult necklace (or)': 1
-		}),
-		outputItems: resolveNameBank({
-			'Occult necklace': 1,
-			'Occult ornament kit': 1
-		}),
-		noCl: true
-	}
-];
-
 const hunterClothing: Createable[] = [
 	{
 		name: 'Polar camouflage gear',
@@ -486,529 +222,6 @@ const hunterClothing: Createable[] = [
 		inputItems: resolveNameBank({ 'Dark kebbit fur': 2 }),
 		outputItems: resolveNameBank({ 'Gloves of silence': 1 }),
 		GPCost: 600
-	}
-];
-
-const slayerCreatables: Createable[] = [
-	{
-		name: 'Dark totem',
-		inputItems: resolveNameBank({
-			'Dark totem base': 1,
-			'Dark totem middle': 1,
-			'Dark totem top': 1
-		}),
-		outputItems: resolveNameBank({ 'Dark totem': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Dragon hunter lance',
-		inputItems: resolveNameBank({
-			"Hydra's claw": 1,
-			'Zamorakian hasta': 1
-		}),
-		outputItems: resolveNameBank({ 'Dragon hunter lance': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Ferocious gloves',
-		inputItems: resolveNameBank({
-			'Hydra leather': 1
-		}),
-		outputItems: resolveNameBank({ 'Ferocious gloves': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Uncut zenyte',
-		inputItems: resolveNameBank({
-			Onyx: 1,
-			'Zenyte shard': 1
-		}),
-		outputItems: resolveNameBank({ 'Uncut zenyte': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Neitiznot faceguard',
-		inputItems: resolveNameBank({
-			'Basilisk jaw': 1,
-			'Helm of neitiznot': 1
-		}),
-		outputItems: resolveNameBank({ 'Neitiznot faceguard': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Arclight',
-		inputItems: resolveNameBank({
-			Darklight: 1,
-			'Ancient shard': 3
-		}),
-		outputItems: resolveNameBank({ Arclight: 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Boots of brimstone',
-		inputItems: resolveNameBank({
-			"Drake's claw": 1,
-			'Boots of stone': 1
-		}),
-		outputItems: resolveNameBank({ 'Boots of brimstone': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Devout boots',
-		inputItems: resolveNameBank({
-			"Drake's tooth": 1,
-			'Holy sandals': 1
-		}),
-		outputItems: resolveNameBank({ 'Devout boots': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Uncharged toxic trident',
-		inputItems: resolveNameBank({
-			'Magic fang': 1,
-			'Uncharged trident': 1
-		}),
-		outputItems: resolveNameBank({ 'Uncharged toxic trident': 1 }),
-		GPCost: 0
-	},
-	{
-		name: "Bryophyta's staff",
-		inputItems: resolveNameBank({
-			'Nature rune': 1000,
-			"Bryophyta's essence": 1,
-			Battlestaff: 1
-		}),
-		outputItems: resolveNameBank({ "Bryophyta's staff": 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Toxic staff (uncharged)',
-		inputItems: resolveNameBank({
-			'Magic fang': 1,
-			'Staff of the dead': 1
-		}),
-		outputItems: resolveNameBank({ 'Toxic staff (uncharged)': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Abyssal tentacle',
-		inputItems: resolveNameBank({
-			'Kraken tentacle': 1,
-			'Abyssal whip': 1
-		}),
-		outputItems: resolveNameBank({ 'Abyssal tentacle': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Brimstone ring',
-		inputItems: resolveNameBank({
-			"Hydra's eye": 1,
-			"Hydra's fang": 1,
-			"Hydra's heart": 1
-		}),
-		outputItems: resolveNameBank({ 'Brimstone ring': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Guardian boots',
-		inputItems: resolveNameBank({
-			'Black tourmaline core': 1,
-			'Bandos boots': 1
-		}),
-		outputItems: resolveNameBank({ 'Guardian boots': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Abyssal bludgeon',
-		inputItems: resolveNameBank({
-			'Bludgeon claw': 1,
-			'Bludgeon spine': 1,
-			'Bludgeon axon': 1
-		}),
-		outputItems: resolveNameBank({ 'Abyssal bludgeon': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Uncharge black mask',
-		inputItems: resolveNameBank({
-			'Black mask (10)': 1
-		}),
-		outputItems: resolveNameBank({ 'Black mask': 1 }),
-		GPCost: 0
-	},
-	{
-		name: 'Slayer ring (8)',
-		inputItems: resolveNameBank({
-			'Gold bar': 1,
-			'Enchanted gem': 1
-		}),
-		outputItems: resolveNameBank({ 'Slayer ring (8)': 1 }),
-		GPCost: 0,
-		requiredSkills: {
-			crafting: 75
-		},
-		requiredSlayerUnlocks: [SlayerTaskUnlocksEnum.RingBling]
-	},
-	{
-		name: 'Slayer ring (eternal)',
-		inputItems: resolveNameBank({
-			'Gold bar': 1,
-			'Eternal gem': 1
-		}),
-		outputItems: resolveNameBank({ 'Slayer ring (eternal)': 1 }),
-		GPCost: 0,
-		requiredSkills: {
-			crafting: 75
-		},
-		requiredSlayerUnlocks: [SlayerTaskUnlocksEnum.RingBling]
-	},
-	{
-		name: 'Slayer helmet',
-		inputItems: resolveNameBank({
-			'Black mask': 1,
-			Earmuffs: 1,
-			Facemask: 1,
-			'Nose peg': 1,
-			'Spiny helmet': 1,
-			'Enchanted gem': 1
-		}),
-		requiredSkills: {
-			crafting: 55
-		},
-		outputItems: resolveNameBank({ 'Slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [SlayerTaskUnlocksEnum.MalevolentMasquerade]
-	},
-	{
-		name: 'Slayer helmet (i)',
-		inputItems: resolveNameBank({
-			'Black mask (i)': 1,
-			Earmuffs: 1,
-			Facemask: 1,
-			'Nose peg': 1,
-			'Spiny helmet': 1,
-			'Enchanted gem': 1
-		}),
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSkills: {
-			crafting: 55
-		},
-		requiredSlayerUnlocks: [SlayerTaskUnlocksEnum.MalevolentMasquerade]
-	},
-	{
-		name: 'Revert slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1 }),
-		outputItems: resolveNameBank({
-			'Black mask': 1,
-			Earmuffs: 1,
-			Facemask: 1,
-			'Nose peg': 1,
-			'Spiny helmet': 1,
-			'Enchanted gem': 1
-		}),
-		requiredSkills: {
-			crafting: 55
-		},
-		GPCost: 0,
-		requiredSlayerUnlocks: [SlayerTaskUnlocksEnum.MalevolentMasquerade]
-	},
-	{
-		name: 'Revert slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1 }),
-		outputItems: resolveNameBank({
-			'Black mask (i)': 1,
-			Earmuffs: 1,
-			Facemask: 1,
-			'Nose peg': 1,
-			'Spiny helmet': 1,
-			'Enchanted gem': 1
-		}),
-		requiredSkills: {
-			crafting: 55
-		},
-		GPCost: 0,
-		requiredSlayerUnlocks: [SlayerTaskUnlocksEnum.MalevolentMasquerade]
-	},
-	{
-		name: 'Black slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1, 'Kbd heads': 1 }),
-		outputItems: resolveNameBank({ 'Black slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KingBlackBonnet
-		]
-	},
-	{
-		name: 'Black slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Kbd heads': 1 }),
-		outputItems: resolveNameBank({ 'Black slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KingBlackBonnet
-		]
-	},
-	{
-		name: 'Revert black slayer helmet',
-		outputItems: resolveNameBank({ 'Slayer helmet': 1, 'Kbd heads': 1 }),
-		inputItems: resolveNameBank({ 'Black slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KingBlackBonnet
-		]
-	},
-	{
-		name: 'Revert black slayer helmet (i)',
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Kbd heads': 1 }),
-		inputItems: resolveNameBank({ 'Black slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KingBlackBonnet
-		]
-	},
-	{
-		name: 'Green slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1, 'Kq head': 1 }),
-		outputItems: resolveNameBank({ 'Green slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KalphiteKhat
-		]
-	},
-	{
-		name: 'Green slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Kq head': 1 }),
-		outputItems: resolveNameBank({ 'Green slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KalphiteKhat
-		]
-	},
-	{
-		name: 'Revert green slayer helmet',
-		outputItems: resolveNameBank({ 'Slayer helmet': 1, 'Kq head': 1 }),
-		inputItems: resolveNameBank({ 'Green slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KalphiteKhat
-		]
-	},
-	{
-		name: 'Revert green slayer helmet (i)',
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Kq head': 1 }),
-		inputItems: resolveNameBank({ 'Green slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.KalphiteKhat
-		]
-	},
-	{
-		name: 'Red slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1, 'Abyssal head': 1 }),
-		outputItems: resolveNameBank({ 'Red slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UnholyHelmet
-		]
-	},
-	{
-		name: 'Red slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Abyssal head': 1 }),
-		outputItems: resolveNameBank({ 'Red slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UnholyHelmet
-		]
-	},
-	{
-		name: 'Revert red slayer helmet',
-		outputItems: resolveNameBank({ 'Slayer helmet': 1, 'Abyssal head': 1 }),
-		inputItems: resolveNameBank({ 'Red slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UnholyHelmet
-		]
-	},
-	{
-		name: 'Revert red slayer helmet (i)',
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Abyssal head': 1 }),
-		inputItems: resolveNameBank({ 'Red slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UnholyHelmet
-		]
-	},
-	{
-		name: 'Purple slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1, 'Dark claw': 1 }),
-		outputItems: resolveNameBank({ 'Purple slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.DarkMantle
-		]
-	},
-	{
-		name: 'Purple slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Dark claw': 1 }),
-		outputItems: resolveNameBank({ 'Purple slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.DarkMantle
-		]
-	},
-	{
-		name: 'Revert purple slayer helmet',
-		outputItems: resolveNameBank({ 'Slayer helmet': 1, 'Dark claw': 1 }),
-		inputItems: resolveNameBank({ 'Purple slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.DarkMantle
-		]
-	},
-	{
-		name: 'Revert purple slayer helmet (i)',
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Dark claw': 1 }),
-		inputItems: resolveNameBank({ 'Purple slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.DarkMantle
-		]
-	},
-	{
-		name: 'Turquoise slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1, "Vorkath's head": 1 }),
-		outputItems: resolveNameBank({ 'Turquoise slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UndeadHead
-		]
-	},
-	{
-		name: 'Turquoise slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1, "Vorkath's head": 1 }),
-		outputItems: resolveNameBank({ 'Turquoise slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UndeadHead
-		]
-	},
-	{
-		name: 'Revert turquoise slayer helmet',
-		outputItems: resolveNameBank({ 'Slayer helmet': 1, "Vorkath's head": 1 }),
-		inputItems: resolveNameBank({ 'Turquoise slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UndeadHead
-		]
-	},
-	{
-		name: 'Revert turquoise slayer helmet (i)',
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1, "Vorkath's head": 1 }),
-		inputItems: resolveNameBank({ 'Turquoise slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UndeadHead
-		]
-	},
-	{
-		name: 'Hydra slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1, 'Alchemical hydra heads': 1 }),
-		outputItems: resolveNameBank({ 'Hydra slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UseMoreHead
-		]
-	},
-	{
-		name: 'Hydra slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Alchemical hydra heads': 1 }),
-		outputItems: resolveNameBank({ 'Hydra slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UseMoreHead
-		]
-	},
-	{
-		name: 'Revert hydra slayer helmet',
-		outputItems: resolveNameBank({ 'Slayer helmet': 1, 'Alchemical hydra heads': 1 }),
-		inputItems: resolveNameBank({ 'Hydra slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UseMoreHead
-		]
-	},
-	{
-		name: 'Revert hydra slayer helmet (i)',
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Alchemical hydra heads': 1 }),
-		inputItems: resolveNameBank({ 'Hydra slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.UseMoreHead
-		]
-	},
-	{
-		name: 'Twisted slayer helmet',
-		inputItems: resolveNameBank({ 'Slayer helmet': 1, 'Twisted horns': 1 }),
-		outputItems: resolveNameBank({ 'Twisted slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.TwistedVision
-		]
-	},
-	{
-		name: 'Twisted slayer helmet (i)',
-		inputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Twisted horns': 1 }),
-		outputItems: resolveNameBank({ 'Twisted slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.TwistedVision
-		]
-	},
-	{
-		name: 'Revert twisted slayer helmet',
-		outputItems: resolveNameBank({ 'Slayer helmet': 1, 'Twisted horns': 1 }),
-		inputItems: resolveNameBank({ 'Twisted slayer helmet': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.TwistedVision
-		]
-	},
-	{
-		name: 'Revert twisted slayer helmet (i)',
-		outputItems: resolveNameBank({ 'Slayer helmet (i)': 1, 'Twisted horns': 1 }),
-		inputItems: resolveNameBank({ 'Twisted slayer helmet (i)': 1 }),
-		GPCost: 0,
-		requiredSlayerUnlocks: [
-			SlayerTaskUnlocksEnum.MalevolentMasquerade,
-			SlayerTaskUnlocksEnum.TwistedVision
-		]
 	}
 ];
 
@@ -2063,7 +1276,8 @@ const Createables: Createable[] = [
 	...hunterClothing,
 	...twistedAncestral,
 	...metamorphPetCreatables,
-	...slayerCreatables
+	...slayerCreatables,
+	...capeCreatables
 ];
 
 export default Createables;
